implement authorization check; clean up templates
